Move MockHexBoostingClient to common
